progress on cycle, hook multi block

+++ code/modules/shuttles/dock/subtypes/ephemeral.dm
@@ -10,7 +10,8 @@
 	create_bounding_box_area = FALSE
 
 /obj/shuttle_dock/ephemeral/on_shuttle_departed(datum/shuttle/shuttle)
-	#warn make sure it doesn't delete if the shuttle is coming right backbaglist
+	if(inbound)
+		return
 	qdel(src)
 
 /**
